+++ hahomematic/platforms/custom/light.py
@@ -525,7 +525,7 @@ class CeIpRGBWLight(CeDimmer):
             and self.channel_no in (3, 4)
         ):
             return EntityUsage.NO_CREATE
-        return self._usage
+        return self._get_entity_usage()
 
     @value_property
     def effects(self) -> tuple[str, ...] | None:
@@ -1117,7 +1117,7 @@ DEVICES: Mapping[str, CustomConfig | tuple[CustomConfig, ...]] = {
     "HMW-LC-Dim1L-DR": CustomConfig(make_ce_func=make_rf_dimmer, channels=(3,)),
     "HSS-DX": CustomConfig(make_ce_func=make_rf_dimmer, channels=(1,)),
     "HmIP-DRG-DALI": CustomConfig(
-        make_ce_func=make_ip_drg_dali_light, channels=tuple(range(1, 49, 1))
+        make_ce_func=make_ip_drg_dali_light, channels=tuple(range(0, 49, 1))
     ),
     "HmIP-BDT": CustomConfig(make_ce_func=make_ip_dimmer, channels=(3,)),
     "HmIP-BSL": CustomConfig(make_ce_func=make_ip_fixed_color_light, channels=(7, 11)),